Add SSR Support
Added `jsdom` as a dependency.

In parser-markup.js, `document` is accessed to create a frame to render HTML in. Since `document` doesn't exist in Node, this causes SSR builds to throw a reference error. With this change, we check for the existence of `document` and polyfill it using `jsdom` if it doesn't exist.